# apps/backend/openapi/v1.yaml
...
openapi: 3.1.1
info:
  title: Argentum API
  # A date rather than a semver, matching what `GET /v1/me` reports as
  # `api_version` — "you are on 2026-07-30" locates a deploy, where "1.0.0"
  # locates nothing. A test pins this equal to `apiv1.Version`.
  version: '2026-07-30'
  summary: Ask for a report or an answer over HTTP.
  description: |
    Argentum turns a question about a company's own data into an answer or a
    branded document. This is the machine-facing half: the same agent the
    dashboard drives, reachable with an API key.

    ## The two products

    **Reports.** `POST /v1/reports/render` takes a document spec and returns a
    file — deterministic, sub-second, no LLM. `POST /v1/reports` takes a prompt
    and runs a real agent turn that ends by generating one.

    **Chat.** `POST /v1/chat` takes a question and returns an answer, either
    streamed over server-sent events or waited for on the connection.

    ## The contract every route inherits

    * **Authentication** is `Authorization: Bearer arg_…`. A dashboard session
      is refused here, and a key is refused on the dashboard's own `/api`
      routes: they are different authorities.
    * **Scopes** are fixed when a key is minted and cannot be changed. Each
      operation below names the one it requires under `x-argentum-scope`.
    * **Every response carries `X-Request-Id`**, including failures. Quote it
      in a support conversation.
...
      `error.message` is for a human.
...

    ## Streams

...

...

...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /v1/agents:
    get:
      operationId: listAgents
      tags: [Identity]
      summary: The agents this workspace has, and which one answers by default
      description: |
        A workspace can keep several agents — Finance, Ops, Support — each with
        its own persona, its own tools and its own databases. This is the list,
        and the `id` of a row here is what goes in `agent_id` on
        `POST /v1/chat` and `POST /v1/reports`.

        It needs no scope, on the same terms as `GET /v1/me`: a key that can
        send a turn has to be able to discover what to send it as.

        `is_default` marks the one a call with no `agent_id` runs as. A row with
        `enabled: false` has been switched off by an admin — it stays in the
        list so a call that started failing has a visible reason, but naming it
        answers `404 agent_not_found` like any other id this workspace cannot
        use.

        A key an admin limited to named agents lists **those agents and no
        others**, because naming any other is the same `404`. If the default is
        not among them, no row carries `is_default: true`, and a call on that key
        has to name its `agent_id`. A key with no list — every key minted before
        the list existed — sees the whole roster, restricted agents included: a
        key is not a person, and the grants an admin gives people do not apply
        to it.

        The whole roster arrives in one page. `has_more` is always `false`; the
        envelope is the one every other `/v1` list uses so a client needs no
        special case for this one.
      x-argentum-scope: null
      responses:
        '200':
          description: The roster, default first.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v1/chat:
    post:
      operationId: sendChat
      tags: [Chat]
      summary: Ask a question
      description: |
        **`Accept` decides the door.** `text/event-stream` streams the turn as
        it is written; anything else blocks until the turn answers.

        A client disconnect never cancels a turn: the agent runs in the worker
        and this route is a reader of the same bus the dashboard reads. Hanging
        up ends the read, the turn finishes, the answer persists, and
        `GET /v1/threads/{id}/messages` collects it.

        The synchronous door answers **504** when the turn outruns the wait —
        not when the turn fails. That response carries `in_flight.thread_id`,
        and it **keeps its idempotency key**, so the retry a 504 invites replays
        instead of starting a second billed turn. Attach to
        `GET /v1/threads/{id}/events` instead.
      x-argentum-scope: write:chat
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - $ref: '#/components/parameters/LastEventID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatRequest' }
            example:
              message: What were total sales last month?
              user_ref: u_42
      responses:
        '200':
          description: |
            The completed turn, or — with `Accept: text/event-stream` — the
            stream that produces it.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Turn' }
            text/event-stream:
              schema: { $ref: '#/components/schemas/ChatEvent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/BudgetExhausted' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
        '504':
          description: |
            The **wait** ran out, not the turn. It is still running and still
            being billed; collect it from the stream rather than asking again.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PendingTurn' }
...
  /v1/threads/{id}/messages:
    get:
      operationId: listMessages
      tags: [Chat]
      summary: Read a transcript
      description: Oldest first, cursor-paginated.
      x-argentum-scope: read:threads
      parameters:
        - $ref: '#/components/parameters/ThreadID'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/UserRefGuard'
      responses:
        '200':
          description: One page of messages.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessagePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer arg_<prefix>_<secret>`, minted in the dashboard
        under Settings → API Keys. The secret is shown once.

        **Header only.** The dashboard's own token may travel in a cookie or a
        query parameter because a browser cannot set a header on a WebSocket
        upgrade; neither applies to a machine caller, and both are how a
        credential ends up in an access log or a proxy trace.
  parameters:
    Limit:
      name: limit
      in: query
      description: |
        Page size. A value over the repository's cap is clamped rather than
        refused — a caller asking for 10 000 rows wants as many as they can get,
        and the cursor is what makes the rest reachable.
      required: false
      schema: { type: integer, minimum: 1, examples: [50] }
    Cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from the previous page. Opaque: it encodes the
        `(created_at, id)` of the last row you saw, and hand-building one is a
        `400`, not a silent restart at page one.
      required: false
      schema: { type: string }
...
    ThreadID:
      name: id
      in: path
      required: true
      description: A thread id.
      schema: { type: string, format: uuid }
    UserRefGuard:
      name: user_ref
      in: query
      required: false
      description: |
        When present, the thread must belong to this end user or the route
        answers 404. We cannot authenticate a `user_ref` — a key belongs to the
        company, not to one of its users — but we can hold a caller to the one
        they named, which is what turns "our backend passes the logged-in user's
        id through" into an isolation boundary rather than a convention.
      schema: { type: string }
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        A unique string per logical request; a UUID is the usual choice. Both
        SDKs generate one automatically when you do not supply it.

        A second request under the same key ends one of three ways:

        * the original finished — you get its result again, with
          `Idempotent-Replay: true`, and nothing is billed twice;
        * the original is still running — `409 request_in_flight`, carrying the
          ids of the work you are already waiting on;
        * the body differs from the original's — `409 idempotency_key_reuse`.
          Use a new key for a new request.

        A request that fails forgets its key, so a retry after a 500 is a fresh
        attempt. The one exception is `POST /v1/chat`'s 504, where the request
        failed and the work did not.
      schema: { type: string, maxLength: 255 }
    LastEventID:
      name: Last-Event-ID
      in: header
      required: false
      description: |
        The `id:` of the last frame you received. An SSE client sends this by
        itself on reconnect. The messages after that point are replayed before
        the stream attaches live. A value this API did not issue is a `400`.
      schema: { type: string }
  headers:
    X-Request-Id:
      description: |
        This request's id, on every response including failures. Send your own
        and it is echoed back, provided it is at most 128 characters of
        `[A-Za-z0-9_.:-]`.
      schema: { type: string, examples: ['req_1945841125adc2c7bb618224f5fa88fd'] }
    RateLimit-Limit:
      description: Requests permitted per minute on this key.
      schema: { type: integer }
    RateLimit-Remaining:
      description: Whole requests left in the bucket after this one.
      schema: { type: integer }
    RateLimit-Reset:
      description: Seconds until at least one request is available. `0` means now.
      schema: { type: integer }
    Retry-After:
      description: Seconds to wait before retrying. Both SDKs honour it.
      schema: { type: integer }
    Idempotent-Replay:
      description: |
        `true` when this response replays an earlier request under the same
        `Idempotency-Key` rather than executing a new one.
      schema: { type: string, enum: ['true'] }
...
  responses:
    BadRequest:
      description: |
        The request is malformed. `error.param` names the offending field
        whenever there is one to name.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: |
        No key, or one that is not usable. One message covers malformed,
        unknown, wrong-secret, revoked and expired: a caller holding a broken
        credential has no business learning which of the five it is.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        A valid key without the scope this route requires. Scopes are fixed when
        a key is minted — mint a new one.

        On `POST /v1/chat` and `POST /v1/reports` it is also
        `agent_not_allowed`: the key is limited to named agents, and the turn
        would run as another. Name one of the key's agents in `agent_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No such resource **for this company**. A cross-tenant id and a
        nonexistent id answer identically; anything else is an existence oracle
        over another company's data.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Two shapes, distinguished by `error.code`. `request_in_flight` carries
        `in_flight` with the ids of the work already running;
        `idempotency_key_reuse` means the same key arrived with a different
        body.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConflictError' }
    BudgetExhausted:
      description: |
        The tenant is out of credit. Typed rather than a 500 on purpose: a
        programmatic caller retries a 500, and retrying a turn nobody can pay
        for is a loop that stops only when someone notices.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: The request body is over this deployment's cap.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests on this key.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: |
        Our fault. Quote `error.request_id` — it is the string that finds this
        attempt in our logs.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Either the whole public API is switched off (`api_disabled`, with
        `Retry-After`) or one subsystem this route needs is not configured on
        this deployment. The routes exist and are coming back, which is why this
        is not a 404.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      description: |
        Every failure under `/v1` has this shape. A bare `{"error": "…"}` string
        anywhere here is a defect.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorDetail' }
      additionalProperties: false
...
    AgentPage:
      type: object
      description: |
        The whole roster, in the envelope every `/v1` list uses. `has_more` is
        always `false` — a roster is small and there is nothing to page — but the
        shape is the shared one so no client special-cases this route.
      required: [data, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Agent' }
        has_more: { type: boolean }
        next_cursor: { type: string }
      additionalProperties: false
...
    ChatRequest:
      type: object
      description: |
        Send `user_ref` **or** `thread_id`. `user_ref` keys the thread when there
        is no `thread_id`, and is what makes the spend attributable.
      required: [message]
      properties:
        message:
          type: string
          description: The question you want answered.
        thread_id:
          type: string
          format: uuid
          description: Continue a conversation you are already tracking.
        user_ref:
          type: string
          description: Your own identifier for the person asking.
        agent_id:
          type: string
          format: uuid
          description: |
            Which of the workspace's agents answers. Omit it for the default —
            which is what every call written before this field existed keeps
            meaning. List them with `GET /v1/agents`.

            It belongs to the conversation rather than to the turn. Sent with a
            `thread_id`, it has to agree with what that conversation already
            runs as, or the call is refused with `agent_mismatch` — a pick that
            was silently ignored would leave a client believing it had switched
            agents while every answer still came from the old one. Sent with a
            `user_ref` whose newest conversation runs as a different agent, a
            new conversation starts, for the same reason a topic change forks
            one.

            An id this workspace does not have — unknown, deleted, disabled, or
            another tenant's — is `404 agent_not_found`, never `403`. A 403
            would confirm the row exists to someone guessing uuids.

            A key an admin limited to named agents answers the same `404` for
            an agent outside its list. It answers `403 agent_not_allowed` when
            the turn would run as an agent the caller did not name: a
            `thread_id` whose conversation runs as one outside the list, or a
            call with no `agent_id` when the workspace default is not on it. The
            fix for both is to send one of the key's agents in `agent_id`. A key
            with no list reaches every agent.
      additionalProperties: false
    Turn:
      type: object
      description: |
        One question and its answer. The same shape arrives as the `final` SSE
        frame, so a caller who starts on the synchronous door and moves to the
        stream does not rewrite their parser.
      required: [object, thread_id, message]
      properties:
        object: { type: string, const: turn }
        thread_id: { type: string, format: uuid }
        run_id:
          type: string
          description: |
            Identifies this turn within the thread. It is the id of the question
            that started it.
        message: { $ref: '#/components/schemas/Message' }
        usage: { $ref: '#/components/schemas/Usage' }
        request_id: { type: string }
      additionalProperties: false
...
    Message:
      type: object
      description: |
        One turn in a transcript. Token counts and latency are deliberately not
        here: they are per-message zeros for every streamed turn, and a field
        that is honestly zero most of the time is a field integrators report as
        a bug.
      required: [object, role, content]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: message }
        role:
          type: string
          enum: [user, assistant, system]
        content: { type: string }
        created_at: { type: string, format: date-time }
      additionalProperties: false
    MessagePage:
      type: object
      required: [data, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Message' }
        has_more: { type: boolean }
        next_cursor: { type: string }
      additionalProperties: false
...
    PendingTurn:
      type: object
      description: |
        The 504 body. The field names match the `409 request_in_flight` on
        purpose: both mean "the work is still running and here is how to find
        it", and a caller should be able to parse them with one branch.
      required: [error, in_flight]
      properties:
        error: { $ref: '#/components/schemas/ErrorDetail' }
        in_flight:
          type: object
          required: [thread_id, started_at]
          properties:
            thread_id: { type: string, format: uuid }
            run_id: { type: string }
            started_at: { type: string, format: date-time }
          additionalProperties: false
      additionalProperties: false
    ChatEvent:
      description: |
        One frame of a chat stream, as the `data:` of an SSE event whose
        `event:` is the `type` below. The names are the dashboard's, unchanged,
        because one worker publishes both surfaces.

        `final` and `error` are terminal. `message` appears only on a resumed
        stream, replaying what a disconnect missed. Only `message` and `final`
        carry an `id:` — they are the frames backed by something persisted.

        Treat an unknown event name as ignorable; that is what lets this list
        grow without breaking you.
      oneOf:
        - $ref: '#/components/schemas/ChatEventStarted'
        - $ref: '#/components/schemas/ChatEventDelta'
        - $ref: '#/components/schemas/ChatEventThinking'
        - $ref: '#/components/schemas/ChatEventTool'
        - $ref: '#/components/schemas/ChatEventMessage'
        - $ref: '#/components/schemas/ChatEventError'
        - $ref: '#/components/schemas/ChatEventFinal'
    ChatEventStarted:
      type: object
      title: started
      description: The worker picked the turn up.
      required: [thread_id, at]
      properties:
        thread_id: { type: string, format: uuid }
        run_id: { type: string }
        at: { type: string, format: date-time }
      additionalProperties: false
...